requirements.txt
Flask>=2.2
pyperclip>=1.8.2
winotify>=1.1.0
pystray>=0.19.5
Pillow>=9.0.0